+++ components/AppFooter.vue
@@ -2,33 +2,33 @@
   <footer class="font-mono absolute z-20 bottom-0 w-screen bg-gray-900">
     <section class="py-8 px-3">
       <nav class="flex gap-20 max-w-5xl my-0 mx-auto">
-        <nuxt-link
+        <NuxtLink
           to="/"
           class="flex align-center text-gray-100 hover:text-yellow-400 focus:text-yellow-400 transition-all duration-200"
           ><img class="h-6 mr-3" src="/st-logo.png" /><span
             >Simply Tiling</span
-          ></nuxt-link
+          ></NuxtLink
         >
         <ul>
           <li class="pb-3">
-            <nuxt-link
+            <NuxtLink
               to="/services"
               class="text-gray-100 hover:text-yellow-400 focus:text-yellow-400 transition-all duration-200"
-              >Our Services</nuxt-link
+              >Our Services</NuxtLink
             >
           </li>
           <li class="pb-3">
-            <nuxt-link
+            <NuxtLink
               to="/gallery"
               class="text-gray-100 hover:text-yellow-400 focus:text-yellow-400 transition-all duration-200"
-              >Gallery</nuxt-link
+              >Gallery</NuxtLink
             >
           </li>
           <li>
-            <nuxt-link
+            <NuxtLink
               to="/get-a-quote"
               class="text-gray-100 hover:text-yellow-400 focus:text-yellow-400 transition-all duration-200"
-              >Get A Quote</nuxt-link
+              >Get A Quote</NuxtLink
             >
           </li>
         </ul>